HDX-9781 Move the admin2 creation/tests
