+++ src/hapi_schema/db_admin2.py
@@ -72,23 +72,3 @@ view_params_admin2 = ViewParams(
         ),
     ),
 )
-
-
-class DBAdmin2VAT(Base):
-    __tablename__ = "admin2_vat"
-    id: Mapped[int] = mapped_column(Integer, primary_key=True)
-    admin1_ref: Mapped[int] = mapped_column(Integer)
-    code: Mapped[str] = mapped_column(String(128), index=True)
-    name: Mapped[str] = mapped_column(String(512), index=True)
-    is_unspecified: Mapped[bool] = mapped_column(Boolean)
-    reference_period_start: Mapped[datetime] = mapped_column(
-        DateTime, index=True
-    )
-    reference_period_end: Mapped[datetime] = mapped_column(
-        DateTime, index=True
-    )
-    admin1_code: Mapped[str] = mapped_column(String(128), index=True)
-    admin1_name: Mapped[str] = mapped_column(String(512), index=True)
-    admin1_is_unspecified: Mapped[bool] = mapped_column(Boolean)
-    location_code: Mapped[str] = mapped_column(String(128), index=True)
-    location_name: Mapped[str] = mapped_column(String(512), index=True)
