+++ tests/test_views_as_tables.py
@@ -1,4 +1,5 @@
 from hapi_schema.db_admin1 import view_params_admin1
+from hapi_schema.db_admin2 import view_params_admin2
 
 
 def test_admin1_vat(run_indexes_test, run_columns_test, run_primary_keys_test):
@@ -15,3 +16,19 @@ def test_admin1_vat(run_indexes_test, run_columns_test, run_primary_keys_test):
     run_columns_test("admin1_vat", "admin1_view", view_params_admin1)
     run_indexes_test("admin1_vat", expected_indexes)
     run_primary_keys_test("admin1_vat", expected_primary_keys)
+
+
+def test_admin2_vat(run_indexes_test, run_columns_test, run_primary_keys_test):
+    """Check that admin2_vat is correct - columns match, expected indexes present"""
+    expected_primary_keys = ["id"]
+    expected_indexes = [
+        "code",
+        "name",
+        "reference_period_start",
+        "reference_period_end",
+        "location_code",
+        "location_name",
+    ]
+    run_columns_test("admin2_vat", "admin2_view", view_params_admin2)
+    run_indexes_test("admin2_vat", expected_indexes)
+    run_primary_keys_test("admin2_vat", expected_primary_keys)
